shopify-insights/app/scraper.py
import httpx
from bs4 import BeautifulSoup
import re
from .models import Product
from .llm_processor import structure_with_llm

class ShopifyScraper:
    def __init__(self, base_url: str):
        self.base_url = base_url.rstrip('/')
        self.session = httpx.AsyncClient(timeout=30.0)
    
    async def get_products(self):
        response = await self.session.get(f"{self.base_url}/products.json")
        response.raise_for_status()
        return [
            Product(
                id=str(p['id']),
                title=p['title'],
                handle=p['handle'],
                description=p.get('body_html'),
                price=p['variants'][0]['price'],
                image_url=p['images'][0]['src'] if p['images'] else None
            ) for p in response.json().get('products', [])
        ]
    
    async def get_hero_products(self):
        # Scrape homepage for featured products
        return []
    
    async def get_policy(self, policy_type: str):
        # Find and scrape policy pages
        return ""
    
    async def get_faqs(self):
        # Find FAQ page and process with LLM
        faq_page = await self.find_page(['FAQ', 'FAQs', 'Help'])
        if faq_page:
            structured_faqs = await structure_with_llm(faq_page, "faqs")
            return structured_faqs
        return []
    
    async def get_social_handles(self):
        # Scrape for social media handles
        return {}
    
    async def get_contact_details(self):
        # Scrape for contact details
        return {}
    
    async def get_about_brand(self):
        # Scrape for about brand section
        return ""
    
    async def get_important_links(self):
        # Scrape for important links
        return {}
    
    async def find_page(self, keywords):
        # Helper to find a page by keywords
        return "" 