shopify-insights/requirements.txt
fastapi
uvicorn
httpx
beautifulsoup4
sqlalchemy
pymysql
pydantic
pydantic-settings
google-generativeai 